requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
chardet==3.0.4
discord.py==1.6.0
idna==3.1
multidict==4.7.6
Pillow==8.1.2
python-dotenv==0.17.1
soupsieve==2.0.1
typing-extensions==3.7.4.3
urllib3==1.25.10
yarl==1.6.3
youtube-dl==2021.3.14
